+++ helper_field/widget.js
@@ -5,7 +5,7 @@
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
  * @since    0.1.0
- * @version  0.1.0
+ * @version  0.2.6
  */
 const WidgetField = Function.inherits('Alchemy.Field.Schema', function Widget(schema, name, options) {
 
@@ -17,18 +17,39 @@ const WidgetField = Function.inherits('Alchemy.Field.Schema', function Widget(sc
 		options.type = 'text';
 	}
 
+	// Already set the options (they'll be changed by the super call too though)
+	this.options = options;
+
 	// A custom schema should NOT be passed to this class, this class uses
 	// a fixed schema that should not be altered.
 	// But because that's exactly what happens when cloning (like preparing
 	// the data to be sent to Hawkejs) we have to allow it anyway
 	if (!options.schema) {
-		let WidgetClass = Classes.Alchemy.Widget.Widget.getMember(options.type),
+		let WidgetClass = this.widget_class,
 		    sub_schema = WidgetClass.schema.clone();
 
 		options.schema = sub_schema;
 	}
 
 	Widget.super.call(this, schema, name, options);
+});
+
+/**
+ * Get the constructor of the widget class
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.6
+ * @version  0.2.6
+ *
+ * @type   {Function}
+ */
+WidgetField.setProperty(function widget_class() {
+
+	if (!this.options?.type) {
+		return;
+	}
+
+	return Classes.Alchemy.Widget.Widget.getMember(this.options.type)
 });
 
 /**
@@ -74,4 +95,32 @@ WidgetField.setMethod(function toDry() {
 	};
 
 	return {value};
+});
+
+/**
+ * See if the given value is considered not-empty for this field
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.6
+ * @version  0.2.6
+ *
+ * @param    {Mixed}   value
+ *
+ * @return   {Boolean}
+ */
+WidgetField.setMethod(function valueHasContent(value) {
+
+	if (!value) {
+		return false;
+	}
+
+	let constructor = this.widget_class;
+
+	if (!constructor) {
+		return true;
+	}
+
+	let instance = new constructor();
+
+	return instance.valueHasContent(value);
 });